+++ CSGO/CSGO/Source/Program/pattern.h
@@ -4,7 +4,7 @@ class CPatternScanner
 {
 private:
 	MODULEENTRY32 getModule(char *module);
-	uintptr_t findPattern(char* base, unsigned int size, char* pattern, char *mask);
+	uintptr_t findPattern(char* base, unsigned long size, char* pattern, char *mask);
 
 public:
 	uintptr_t findPatternEx(HANDLE hProcess, char *module, char *pattern, char *mask);
